+++ CHANGELOG.md
@@ -27,7 +27,7 @@ Classify the change according to the following categories:
 
 ## pv-resilience
 ### Added
-- Optional input **PV.outage_production_reduction_fraction** to reduce PV production during modeled outages. Only applies with multiple outage modeling using inputs outage_start_time_steps and outage_durations.
+- Optional input **PV.outage_production_fraction** to reduce PV production during modeled outages. Only applies with multiple outage modeling using inputs outage_start_time_steps and outage_durations.
 
 ## v0.61.1
 ### Fixed